requirements.txt
pandas
openai
psycopg
python-dotenv
timescale-vector
instructor
anthropic
pymupdf
google-api-python-client
google-auth-oauthlib
langchain
pydantic>=2.0.0
reportlab
python-docx